backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0

SQLAlchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1

pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

python-multipart==0.0.20

httpx==0.28.1

